Bug 1501005 - refactor code out of utils into base
